requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
astunparse==1.6.3
attrs==25.4.0
Automat==25.4.16
bcrypt==5.0.0
blinker==1.4
blosc2==2.7.1
Box2D==2.3.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.3.1
cloudpickle==3.1.2
comprl @ git+https://github.com/martius-lab/teamproject-competition-server.git@afb8f28ce94a4d3e706e4637d93450b93c16ccb4#subdirectory=comprl
constantly==23.10.4
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.11
dbus-python==1.2.18
distro==1.7.0
docstring_parser==0.17.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.64.1
gymnasium==1.2.3
h5py==3.11.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib-metadata==4.6.4
Incremental==24.11.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
keras==3.4.1
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
ndindex==1.9.2
networkx==3.2.1
nibabel==5.2.1
numcodecs==0.13.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.2.0
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openskill==6.0.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.2.0
platformdirs==4.5.1
protobuf==4.25.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
Pyro4==4.82
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.3
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
sentry-sdk==2.52.0
serpent==1.41
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
SQLAlchemy==2.0.46
swig==4.4.1
sympy==1.12
tables==3.10.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
triton==3.0.0
Twisted==24.11.0
typeguard==4.4.4
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.3
tzdata==2024.2
urllib3==2.2.2
variconf==1.0.1
wadllib==1.3.6
wandb==0.24.2
Werkzeug==3.0.3
wrapt==1.16.0
zarr==2.18.3
zipp==1.0.0
zope.interface==8.2
